requirements.txt
flask
numpy
pymongo
gunicorn
python-dotenv